feat: Replace DNS record observer with retry policy
This makes use of the Failsafe library instead of a custom DNS record observer
to wait until a challenge completes after inserting the DNS record.

Waiting for the record update does not give us anything in particular in this
case.

The default retry policy tries with increasing backoff delays for up to 10
minutes. In the future this may become configurable.