feat all: Reinstate DnsRecordObserer
Once an ACME challenge reaches 'invalid' state (e.g. if the DNS record
is not yet present) it seems to be impossible to retry the challenge.

Therefore the controller now waits for the records to update again.

In addition the DnsRecordObserver will check whether the returned record
contains the challenge string, instead of checking whether the string
matches entirely.
This is due to the discrepancies with using quotes.