feat Reconciliation: New process model
The Letsencrypt Controller now has a slightly updated process model.

The Main class starts up a namespace manager that keeps track of all
monitored namespaces and reacts to added / removed namespaces.

For every namespace a new thread is started which contains that
namespace's reconciliation loop. This loop runs every minute, checks the
annotations on each service and attempts to reconcile differences by
requesting new certificates.

The service managers keep track of in-progress certificate requests in a
ConcurrentSet, however this is only intended as a weak protection
against parallel runs. If an (unlikely) race-condition leads to a
parallel run it will still be eventually reconciled, it may just have to
time out once.

Old certificates are never removed automatically.

Fixes #5