feat renewals: Check & renew existing certificates
Implement renewal of certificates a day or so before they expire.

Certificates returned by Letsencrypt have a relatively short lifetime.
We store the expiry date in an annotation and check it as a part of the
normal reconciliation loop.

If a certificate is about to expire we renew it and update the cluster
secret.

Fixes #1