fix CloudDNS: Add extra waiting time after update
Sometimes Cloud DNS updates have not finished when the update status is
set to "DONE". This adds an additional one-minute wait timer before
continuing with the challenge validation.

In most cases this seems to work fine. If it fails a few times it will
be corrected in the next runs.

Either way I need to mark this as a separate issue and figure out if
there is a saner way of dealing with it.