feat acme: Add support for SAN certificates
Add support for requesting a certificate for multiple domains at the
same time (SAN certificate).

This is implemented via some additional syntax for the current request
annotation on k8s services.

If a service is a JSON array (checked by comparing the first character
to '[') it is decoded into a list of domains.

If multiple domains are specified the `acme/secretName` annotation MUST
be set to specify the name of the resulting secret.

Fixes #21
Fixes #22
Fixes #17