feat build: Add Gradle configuration
Adds configuration to build the project with Gradle. Some tests are
currently failing.

Gradle seems to be better supported in Kotlin-land, despite being
significantly more "verbose" than Maven in terms of how many files are
around and how many things need to be configured manually.

I'll stick with it until the Kotlin rewrite is done and then make a
decision.

The current Gradle configuration does not yet build a functional
Docker image.