fix build: Fix usage of Gradle application plugin
The Gradle Script Kotlin examples demonstrate the use of the
application plugin using an application {} configuration block,
however this requires some hidden beta feature to be enabled and it is
not documented how to even do that.

This fixes the application plugin configuration to use a different
syntax that I found far down in a Github CHANGELOG, and this one works.