fix route53: Don't select private zones for DNS challenges
In certain cases users may have multiple DNS zones for the same domain
configured, for example in the case of a split-brain DNS setup.

Previously the private zone could have been selected for placing the
records, depending on the order of the result returned from the AWS
API.

This commit will filter out all private zones as it never makes sense
to write a challenge response into them.

This fixes issue #61