fix acme: Correctly write out certificate full-chain
The new method for writing out certificate chains that one of the
previous commits adopted always writes out full chains.

Due to the manual concatenation happening in previous code this
resulted in the "fullchain" field always containing the actual
certificate twice.

This commit fixes that, however the content of the "chain" and
"fullchain" fields are now the same as there isn't a convenient method
for writing out the chain only.